refactor(InputView,OutputView): 입력, 출력 기능 분리
